Marionette support for wpt print reftests
Render print reftests to a PDF using the printing machinary, and use
pdf.js from gecko itself to convert the PDF to an image for pixel
comparisons.

For the surrounding reftest machinery, the main change here is that we
convert reftests to work in terms of lists of images rather than a
single image. For normal reftests we only have a single image in the
list; for print reftests we can have several (one per page). This is
implemented in terms of iterators to avoid unnecessary renders when
the test fails on an earlier page.

Differential Revision: https://phabricator.services.mozilla.com/D79081

bugzilla-url: https://bugzilla.mozilla.org/show_bug.cgi?id=1644395
gecko-commit: ceff82dfdd5e5dc91d2fced74c92659443071b89
gecko-integration-branch: autoland
gecko-reviewers: marionette-reviewers, hiro, maja_zf